Rename files and vars
