+++ solana-programs/payment-router/Anchor.toml
@@ -13,5 +13,5 @@ wallet = "id.json"
 
 [scripts]
 test = "yarn run ts-mocha -p ./tsconfig.json -t 1000000 tests/**/*.ts"
-test-create-pda= "yarn run ts-mocha -g 'pda' -p ./tsconfig.json -t 1000000 tests/**/*.ts"
+test-create-payment-router-balance-pda= "yarn run ts-mocha -g 'pda' -p ./tsconfig.json -t 1000000 tests/**/*.ts"
 test-route= "yarn run ts-mocha -g 'route' -p ./tsconfig.json -t 1000000 tests/**/*.ts"
